requirements.txt
yaml
python-box
torch 
pillow
numpy